+++ src/components/Game/GameWerewolvesSide/GameWerewolvesSide.vue
@@ -18,7 +18,7 @@
                     <div class="d-flex">
                         <PlayerThumbnail :player="player" class="ml-1"/>
                         <transition-group name="fade-list" class="d-flex flex-wrap align-items-center flex-grow-1 p-1">
-                            <PlayerAttribute v-for="({ name, source }) in player.attributes" :key="name" :attribute="name" :source="source"/>
+                            <PlayerAttribute v-for="attribute in player.attributes" :key="attribute.name" :attribute="attribute" :player="player"/>
                         </transition-group>
                     </div>
                 </div>