+++ src/components/shared/Game/PlayerAttribute/PlayerAttribute.vue
@@ -1,10 +1,12 @@
 <template>
-    <img v-tooltip="tooltip" alt="Attribute" class="player-attribute" :class="{ 'not-active': !isAttributeActive }" :src="playerAttribute.SVG"/>
+    <img v-tooltip="tooltip" alt="Attribute" class="player-attribute" :class="{ 'not-active': !isAttributeActive || !player.isAlive }"
+         :src="playerAttribute.SVG"/>
 </template>
 
 <script>
 import { mapGetters } from "vuex";
 import { isPlayerAttributeActive } from "@/helpers/functions/Player";
+import Player from "@/classes/Player";
 import sheriffSVG from "@/assets/svg/attributes/sheriff.svg";
 import seenSVG from "@/assets/svg/actions/look.svg";
 import eatenSVG from "@/assets/svg/attributes/eaten.svg";
@@ -17,23 +19,27 @@ import worshipedSVG from "@/assets/svg/attributes/worshiped.svg";
 import bigBadWolfSVG from "@/assets/svg/roles/big-bad-wolf.svg";
 import charmedSVG from "@/assets/svg/attributes/charmed.svg";
 import cantVoteSVG from "@/assets/svg/attributes/cant-vote.svg";
+import endlessSVG from "@/assets/svg/misc/endless.svg";
+import hourglassSVG from "@/assets/svg/misc/hourglass.svg";
+import chronometerSVG from "@/assets/svg/misc/chronometer.svg";
+import deadSVG from "@/assets/svg/attributes/dead.svg";
 
 export default {
     name: "PlayerAttribute",
     props: {
-        attribute: {
-            type: String,
+        player: {
+            type: Player,
             required: true,
         },
-        source: {
-            type: String,
+        attribute: {
+            type: Object,
             required: true,
         },
     },
     // eslint-disable-next-line max-lines-per-function
     data() {
         return {
-            attributes: {
+            attributesMetadata: {
                 "sheriff": {
                     all: {
                         tooltip: this.$t("PlayerAttribute.attributes.sheriff"),
@@ -118,21 +124,57 @@ export default {
     computed: {
         ...mapGetters("game", { game: "game" }),
         playerAttribute() {
-            if (this.attributes[this.attribute] && this.attributes[this.attribute][this.source]) {
-                return this.attributes[this.attribute][this.source];
+            const { name, source } = this.attribute;
+            if (this.attributesMetadata[name] && this.attributesMetadata[name][source]) {
+                return this.attributesMetadata[name][source];
             }
             return {};
         },
         tooltip() {
             return {
-                content: `<div class="mb-2">
-                                ${this.playerAttribute.tooltip}
-                          </div>
-                          <img width="50" alt="Player Attribute" src="${this.playerAttribute.SVG}"/>`,
+                content: `
+                            <div class="d-flex align-items-center">
+                                <img width="30" alt="Player Attribute" src="${this.playerAttribute.SVG}" class="mr-1"/>
+                                <div class="text-center flex-grow-1">${this.playerAttribute.tooltip}</div>
+                            </div>
+                            <hr class="my-1 bg-dark"/>
+                            <div class="d-flex align-items-center">
+                                <img width="30" alt="Player Attribute Remaining Phases" src="${this.tooltipDescription.icon}"
+                                     class="mr-1"/>
+                                <div class="text-center flex-grow-1">${this.tooltipDescription.text}</div>
+                            </div>
+                        `,
+            };
+        },
+        tooltipDescription() {
+            const { remainingPhases, activeAt } = this.attribute;
+            if (!this.player.isAlive) {
+                return {
+                    text: this.$t("PlayerAttribute.attributeInactiveBecauseDead"),
+                    icon: deadSVG,
+                };
+            } else if (!this.isAttributeActive && activeAt) {
+                const description = { icon: chronometerSVG };
+                if (this.game.turn === activeAt.turn) {
+                    description.text = this.$t("PlayerAttribute.attributeWillBeActiveWhenSunRises");
+                } else {
+                    const remainingTurns = activeAt.turn - this.game.turn;
+                    description.text = this.$tc("PlayerAttribute.remainingTurnsToBeActive", remainingTurns, { remainingTurns });
+                }
+                return description;
+            } else if (remainingPhases) {
+                return {
+                    text: this.$tc("PlayerAttribute.remainingPhases", remainingPhases, { remainingPhases }),
+                    icon: hourglassSVG,
+                };
+            }
+            return {
+                text: this.$t("PlayerAttribute.endlessAttribute"),
+                icon: endlessSVG,
             };
         },
         isAttributeActive() {
-            return isPlayerAttributeActive(this.playerAttribute, this.game);
+            return isPlayerAttributeActive(this.attribute, this.game);
         },
     },
 };
@@ -150,6 +192,10 @@ export default {
 
         &:hover {
             border-color: #D4D4D4;
+        }
+
+        &.not-active {
+            filter: grayscale(1);
         }
     }