+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [#98](https://github.com/antoinezanardi/werewolves-assistant-web/issues/98) - Extend player limit.
 * [#109](https://github.com/antoinezanardi/werewolves-assistant-web/issues/109) - Add limits to waking up interval options.
 * [#110](https://github.com/antoinezanardi/werewolves-assistant-web/issues/110) - Better filter for vue-select.
+* [#112](https://github.com/antoinezanardi/werewolves-assistant-web/issues/112) - Add remaining phases and active in for attributes in sides.
 * [#113](https://github.com/antoinezanardi/werewolves-assistant-web/issues/113) - Add an icon, and a tooltip in player thumbnails in sides.
 
 ### 🐛 Bug fixes